+++ packages/form-upload/src/Upload.js
@@ -115,7 +115,7 @@ const Upload = ({
       setFieldError,
       onDeliverySuccess,
       onDeliveryError,
-      isCloud
+      isCloud,
     ]
   );
 
@@ -149,8 +149,24 @@ const Upload = ({
     onFileUpload,
   ]);
 
-  const removeFile = (fileId) => {
-    const newFiles = fieldValue.filter((file) => file.id !== fileId);
+  const validator = useCallback(
+    (file) => {
+      const isDuplicate = fieldValue.some((prev) => prev.file.name === file.name);
+      if (isDuplicate) {
+        return {
+          code: 'duplicate-name',
+          message: 'A file with this name already exists',
+        };
+      }
+
+      return null;
+    },
+    [fieldValue]
+  );
+
+  const removeFile = (fileId, fileIndex) => {
+    const newFiles = fieldValue.filter((file, index) => file.id !== fileId || index !== fileIndex);
+
     if (newFiles.length !== fieldValue.length) {
       setFieldValue(name, newFiles, true);
       const removedFile = fieldValue.find((file) => file.id === fileId);
@@ -192,10 +208,14 @@ const Upload = ({
       const upload = new UploadCore(file, options);
       await upload.generateId();
 
+      const validationResult = validator(file);
+
       if (file.dropRejectionMessage) {
         upload.errorMessage = file.dropRejectionMessage;
       } else if (totalMaxSize && totalSize + newTotalSize + upload.file.size > totalMaxSize) {
         upload.errorMessage = 'Total documents size is too large';
+      } else if (validationResult) {
+        upload.errorMessage = validationResult.message;
       } else {
         upload.start();
         newTotalSize += upload.file.size;
@@ -254,7 +274,13 @@ const Upload = ({
         <Input name={name} style={{ display: 'none' }} />
         <InputGroup disabled={disabled} className={classes}>
           <Suspense fallback={fallback}>
-            <Dropzone onDrop={onDrop} multiple={multiple} maxSize={maxSize} accept={allowedFileTypes}>
+            <Dropzone
+              onDrop={onDrop}
+              multiple={multiple}
+              maxSize={maxSize}
+              accept={allowedFileTypes}
+              validator={validator}
+            >
               {({ getRootProps, getInputProps, isDragActive }) => (
                 <section>
                   <div